Add time table

+++ lib/data/student_time_table.dart
@@ -36,8 +36,8 @@ class Data {
   TeacherId? teacherId;
   TeacherId? subjectId;
   TeacherId? tableId;
-  int? startTime;
-  int? endTime;
+  double? startTime;
+  double? endTime;
   String? weekDay;
   TeacherId? classRoomId;
   TeacherId? standardId;
